+++ CHANGELOG.md
@@ -1,6 +1,28 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## 1.0.0 - 12.02.2020
+## 1.1.2
+
+* [Fixed] Webpack config
+* [Fixed] Security fixes
+
+## 1.1.1
+
+* [Added] Missing directory handling (displays warning about missing directory and processes the other)
+* [Fixed] Option to set `false` for a `scripts` or `styles` path is now working properly
+* [Fixed] isProduction constant is now defined based on cli args, not only NODE_ENV
+
+## 1.1.0
+
+* [Added] Images handling
+* [Changed] Paths configuration
+
+## 1.0.1
+
+* [Added] SourceMaps
+* [Added] Option tu turn off the DependencyExtractionWebpackPlugin
+* [Fixed] RemoveSuperflousAssetsPlugin fixed
+
+## 1.0.0
 
 Initial release